Refractor NavBar and Route config
